+++ SpriteSheetPacker/ScalingVariantWidget.h
@@ -12,7 +12,7 @@ class ScalingVariantWidget : public QWidget
     Q_OBJECT
 
 public:
-    explicit ScalingVariantWidget(QWidget *parent = 0, const QString& name = "", float scale = 1, int maxTextureSize = 2048, bool pow2 = false);
+    explicit ScalingVariantWidget(QWidget *parent = 0, const QString& name = "", float scale = 1, int maxTextureSize = 2048, bool pow2 = false, bool forceSquared = false);
     ~ScalingVariantWidget();
 
     void setRemoveEnabled(bool enable);
@@ -21,9 +21,11 @@ public:
     float   scale();
     int     maxTextureSize();
     bool    pow2();
+    bool    forceSquared();
 
     void    setPow2(bool enable);
     void    setEnabledPow2(bool enable);
+    void    setForceSquared(bool enable);
 
 signals:
     void remove();
@@ -35,6 +37,7 @@ private slots:
     void on_nameLineEdit_textChanged(const QString &arg1);
     void on_maxTextureSizeComboBox_currentTextChanged(const QString &arg1);
     void on_pow2CheckBox_toggled(bool checked);
+    void on_forceSquaredCheckBox_toggled(bool checked);
 
 private:
     Ui::ScalingVariantWidget *ui;